added permission checking and fixed some bugs

+++ src/Terminal.java
@@ -1,6 +1,4 @@
-import java.util.List;
-import java.util.Objects;
-import java.util.Scanner;
+import java.util.*;
 
 public class Terminal
 {
@@ -35,26 +33,43 @@ public class Terminal
         //Do command
         if(Objects.equals(command, "CreateFile"))
         {
-            System.out.println("CREATING FILE...");
             List<String> args = CommandParser.getArgs();
+            if(args.size() < 2){
+                System.out.println("too few arguments, please enter the command again.");
+                return;
+            }
+
+            if(!CapManager.hasPermission(args.get(0), true))
+                return;
+
+            System.out.println("CREATING FILE...");
             DiskManipulator.allocator.createFile(args.get(0), Integer.parseInt(args.get(1)));
         }
         else if (Objects.equals(command, "CreateFolder"))
         {
-            System.out.println("CREATING FOLDER...");
             List<String> args = CommandParser.getArgs();
+            if(!CapManager.hasPermission(args.get(0), true))
+                return;
+
+            System.out.println("CREATING FOLDER...");
             DiskManipulator.allocator.createFolder(args.get(0));
         }
         else if (Objects.equals(command, "DeleteFile"))
         {
-            System.out.println("DELETING FILE...");
             List<String> args = CommandParser.getArgs();
+            if(!CapManager.hasPermission(args.get(0), false))
+                return;
+
+            System.out.println("DELETING FILE...");
             DiskManipulator.allocator.deleteFile(args.get(0));
         }
         else if (Objects.equals(command, "DeleteFolder"))
         {
-            System.out.println("DELETING FOLDER...");
             List<String> args = CommandParser.getArgs();
+            if(!CapManager.hasPermission(args.get(0), false))
+                return;
+
+            System.out.println("DELETING FOLDER...");
             if(!DiskManipulator.allocator.deleteFolder(args.get(0)))
                 System.out.println("Cannot delete folder at "+args.get(0));;
         }
@@ -89,7 +104,6 @@ public class Terminal
         }
         else if(command.equalsIgnoreCase("Grant"))
         {
-            //TODO
             List<String> args = CommandParser.getArgs();
             if(args.size() != 3)
             {
